requirements.txt
einops==0.8.0
ftfy==6.3.1
numpy==1.22.4
Pillow==10.1.0
Pillow==11.0.0
pytz==2023.3.post1
regex==2024.11.6
scikit_learn==1.3.2
scipy==1.14.1
setuptools==67.8.0
torch==1.13.1+cu117
torchvision==0.14.1+cu117
tqdm==4.65.0
